+++ tests/tests2.py
@@ -9,6 +9,7 @@ pd.set_option('display.width', 1000)          # 宽度足够大
 pd.set_option('display.max_columns', None)     # 显示所有列
 pd.set_option('display.max_colwidth', 30)      # 限制单列宽度
 pd.set_option('display.expand_frame_repr', False)  # 禁用折叠
+pd.set_option('display.max_rows', None) 
 
 # 获取项目根目录的绝对路径
 project_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
@@ -33,6 +34,8 @@ client = FMPClient(api_key=API_Key, output_format='pandas')
 print("\n--- Testing Financials ---")
 
 # Test income statement
-print(f"Fetching income statement for {ticker}...")
+# print(f"Fetching income statement for {ticker}...")
 income_statement = client.financials.get_financials(statement='income',symbol=ticker, limit=4)
-print("Income Statement (first row):\n", income_statement)
+# print("Income Statement (first row):\n", income_statement)
+client.financials.get_merged_financials(symbol=ticker,limit=4).to_csv(f'tests/{ticker}.csv')
+# print(client.stocks.historical_price_full(symbol=ticker,period=1))